Test for possible mac crash fix

+++ toast.py
@@ -287,8 +287,9 @@ class PesterToast(QtWidgets.QWidget, DefaultToast):
     @QtCore.pyqtSlot(QtCore.QVariant)
     def updateBottomLeftAnimation(self, value):
         p = QtWidgets.QApplication.desktop().availableGeometry(self).bottomRight()
-        val = float(self.height())/100
-        self.move(p.x()-self.width(), p.y() - (value * val) +1)
+        val = (self.height())/100
+        # Does type casting this to an int have any negative consequences?
+        self.move(int(p.x()-self.width()), int(p.y() - (value * val) +1))
         self.layout().setSpacing(0)
         QtWidgets.QWidget.show(self)
